[deckhouse] module controller
